Add fix for non-integer episode numbers (#1423)

+++ src/en/animepahe/src/eu/kanade/tachiyomi/animeextension/en/animepahe/dto/AnimePaheDto.kt
@@ -41,5 +41,5 @@ data class EpisodeDto(
     val createdAt: String,
     val session: String,
     @SerialName("episode")
-    val episodeNumber: Int,
+    val episodeNumber: Float,
 )
